feat(auto_kms): move tables to orm package
