+++ pkgs/standards/auto_kms/auto_kms/orm/key.py
@@ -1,4 +1,4 @@
-# auto_kms/tables/key.py
+# auto_kms/orm/key.py
 from __future__ import annotations
 import base64
 from enum import Enum
