+++ pkgs/standards/auto_kms/tests/unit/test_key_lifecycle.py
@@ -6,7 +6,7 @@ import pytest
 from fastapi.testclient import TestClient
 from sqlalchemy import select
 from autoapi.v3.tables import Base
-from auto_kms.tables.key_version import KeyVersion
+from auto_kms.orm import KeyVersion
 
 
 @pytest.fixture
