+++ pkgs/standards/auto_kms/tests/unit/test_key_rotate.py
@@ -7,8 +7,8 @@ from fastapi.testclient import TestClient
 from sqlalchemy import select
 
 from autoapi.v3.tables import Base
-from auto_kms.tables.key import Key
-from auto_kms.tables.key_version import KeyVersion
+from auto_kms.orm import Key
+from auto_kms.orm import KeyVersion
 
 
 @pytest.fixture
