+++ pkgs/standards/auto_kms/tests/unit/test_key_wrap_unwrap_schema.py
@@ -1,7 +1,7 @@
 """Test wrap/unwrap endpoint schema validation."""
 
 from autoapi.v3.bindings import bind
-from auto_kms.tables.key import Key
+from auto_kms.orm import Key
 
 
 def test_key_wrap_unwrap_schemas():
